Healing items implemented on test level
healing items are implemented on the test level. Working on getting a health bar sprite and then waiting on the enemy stuff from Kyle in order to finish off the health script.
