+++ Assets/Prefabs/Player.prefab
@@ -14,6 +14,7 @@ GameObject:
   - component: {fileID: 9153042883107652899}
   - component: {fileID: 5495002955294341502}
   - component: {fileID: 745829555252119656}
+  - component: {fileID: 5738996269611469754}
   m_Layer: 0
   m_Name: Player
   m_TagString: Player
@@ -133,6 +134,9 @@ MonoBehaviour:
   moveDir: {x: 0, y: 0}
   isMeleeing: 0
   meleeArea: {fileID: 8507161216502267023}
+  health: 10
+  collectedItem: {fileID: 0}
+  healthText: {fileID: 0}
 --- !u!114 &9153042883107652899
 MonoBehaviour:
   m_ObjectHideFlags: 0
@@ -209,6 +213,19 @@ Rigidbody2D:
   m_SleepingMode: 1
   m_CollisionDetection: 1
   m_Constraints: 0
+--- !u!114 &5738996269611469754
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 833193759461314858}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 283f3bf621cfa4660a6d34a7d3392d94, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  startHealth: 100
 --- !u!1 &2060872873663126492
 GameObject:
   m_ObjectHideFlags: 0
