+++ Assets/Scenes/TestLevel.unity
@@ -176,6 +176,63 @@ PrefabInstance:
     m_AddedGameObjects: []
     m_AddedComponents: []
   m_SourcePrefab: {fileID: 100100000, guid: 25cbabce2aa3344e28f10ecd2ab9f2b6, type: 3}
+--- !u!1001 &545283630
+PrefabInstance:
+  m_ObjectHideFlags: 0
+  serializedVersion: 2
+  m_Modification:
+    serializedVersion: 3
+    m_TransformParent: {fileID: 0}
+    m_Modifications:
+    - target: {fileID: 1278271356574267757, guid: 543bc4a214b03034a9c51d06af7f4930, type: 3}
+      propertyPath: m_LocalPosition.x
+      value: 5.2
+      objectReference: {fileID: 0}
+    - target: {fileID: 1278271356574267757, guid: 543bc4a214b03034a9c51d06af7f4930, type: 3}
+      propertyPath: m_LocalPosition.y
+      value: 22.54
+      objectReference: {fileID: 0}
+    - target: {fileID: 1278271356574267757, guid: 543bc4a214b03034a9c51d06af7f4930, type: 3}
+      propertyPath: m_LocalPosition.z
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 1278271356574267757, guid: 543bc4a214b03034a9c51d06af7f4930, type: 3}
+      propertyPath: m_LocalRotation.w
+      value: 1
+      objectReference: {fileID: 0}
+    - target: {fileID: 1278271356574267757, guid: 543bc4a214b03034a9c51d06af7f4930, type: 3}
+      propertyPath: m_LocalRotation.x
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 1278271356574267757, guid: 543bc4a214b03034a9c51d06af7f4930, type: 3}
+      propertyPath: m_LocalRotation.y
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 1278271356574267757, guid: 543bc4a214b03034a9c51d06af7f4930, type: 3}
+      propertyPath: m_LocalRotation.z
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 1278271356574267757, guid: 543bc4a214b03034a9c51d06af7f4930, type: 3}
+      propertyPath: m_LocalEulerAnglesHint.x
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 1278271356574267757, guid: 543bc4a214b03034a9c51d06af7f4930, type: 3}
+      propertyPath: m_LocalEulerAnglesHint.y
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 1278271356574267757, guid: 543bc4a214b03034a9c51d06af7f4930, type: 3}
+      propertyPath: m_LocalEulerAnglesHint.z
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 6107885248787220764, guid: 543bc4a214b03034a9c51d06af7f4930, type: 3}
+      propertyPath: m_Name
+      value: HealingItem
+      objectReference: {fileID: 0}
+    m_RemovedComponents: []
+    m_RemovedGameObjects: []
+    m_AddedGameObjects: []
+    m_AddedComponents: []
+  m_SourcePrefab: {fileID: 100100000, guid: 543bc4a214b03034a9c51d06af7f4930, type: 3}
 --- !u!1 &604384636
 GameObject:
   m_ObjectHideFlags: 0
@@ -38129,3 +38186,4 @@ SceneRoots:
   - {fileID: 204241158}
   - {fileID: 1368661127}
   - {fileID: 1992607367}
+  - {fileID: 545283630}
