+++ Assets/Scripts/Bryan_Test/GameManager.cs
@@ -81,5 +81,6 @@ public class GameManager : MonoBehaviour
             OnCameraInstantiated?.Invoke();
         }
     }
+    
 
 }
