--- Assets/Scripts/HealingItem.cs
@@ -1,11 +1,0 @@
-using UnityEngine;
-
-public class HealingItem : BaseItem
-{
-    [SerializeField] private int healAmount = 2;
-    public override void Collect(Player player)
-    {
-        player.AddHealth(healAmount);
-        Destroy(gameObject);
-    }
-}
