+++ Assets/Scripts/Item_Scripts.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: c75f48eef4357407d9ee9cc44089c3dc
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
