+++ Assets/Scripts/Player_Scripts.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 1d92f47b539a04324bcb5291bc5c2695
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
