+++ Assets/Scripts/Player_Scripts/PlayerSideView.cs
